Add note in README about mail setup and registration errors
Added a section to the README explaining that if email settings aren't configured or there's no internet connection, the app will throw an error during registration. Clarified that the user account is still created and they can log in directly. Also included a tip to use the log mailer during development to avoid errors when email sending isn't set up.